test(validation): refactor script
We refactor the validation script for better maintainability.
